util/encoding: move TestPrettyPrintValueEncoded into encoding_test pkg
This will be needed to avoid an import cycle in the following commit.

Release note: None
