+++ pkg/util/encoding/BUILD.bazel
@@ -37,6 +37,7 @@ go_test(
         "encoding_test.go",
         "float_test.go",
         "main_test.go",
+        "pretty_print_test.go",
         "printer_test.go",
     ],
     embed = [":encoding"],
